requirements.txt
dot3k
pyusb
